Add class ImgListLayered
The code was adapted to include the minimum requirements to work with
multi layered fits files. For that, there is the image list class
ImgListLayered. Few other functions had to be adapted as well. Further,
several exceptions are included to base class methods to make the code
more understandable for the end user.