Bug Fix: Setting of calibration coefficients
When initializing a `CalibData` object with the `calib_coeffs` argument, they were not set.
`self._calib_coeffs` is initialized as None, so `len(self._calib_coeffs)` will throw an error. However inside __init__ the setting of the `self.calib_coeffs` is inside an try-except block and the error goes unnoticed.
The if statement should fix that but the question remains if the warning is necessary and/or also if an separte object should be avialble which only contains calibration results (which were e.g. obtained from other methods or dummy values for fast visualising of data).