Add option to return DataFrame from get_mean_value
The default behaviour of get_mean_value and get_mean_tseries_rects was not changed. Using the boolean argument return_dataframe, the user can decide if a pandas.DataFrame or a PixelMeanTimeSeries object should be returned. PixelMeanTimeSeriesinherits from pandas.Series but extends its functionality only for some limited cases, the additionals methods are used within the cellcalib module but might not be applicable to the end user's application. Returning a widely spread data type improves the intuitiveness and employability of pyplis.